+++ pydocx/openxml/wordprocessing/run_properties.py
@@ -25,6 +25,17 @@ class RunProperties(XmlModel):
     parent_style = XmlChild(name='rStyle', attrname='val')
     pos = XmlChild(name='position', attrname='val')
     sz = XmlChild(name='sz', attrname='val')
+    clr = XmlChild(name='color', attrname='val')
+
+    @property
+    def color(self):
+        if self.clr is None:
+            return
+        # TODO: When we support background colors, remove FFFFFF check
+        if self.clr == '000000' or self.clr == 'FFFFFF':
+            return
+
+        return self.clr
 
     @property
     def position(self):