+++ pydocx/openxml/markup_compatibility/alternate_content.py
@@ -1,0 +1,14 @@
+# coding: utf-8
+from __future__ import (
+    absolute_import,
+    print_function,
+    unicode_literals,
+)
+
+from pydocx.models import XmlModel, XmlCollection
+from pydocx.openxml.markup_compatibility.fallback import Fallback
+
+
+class AlternateContent(XmlModel):
+    XML_TAG = 'AlternateContent'
+    children = XmlCollection(Fallback)