+++ pydocx/openxml/vml/__init__.py
@@ -1,8 +1,10 @@
 # coding: utf-8
 from pydocx.openxml.vml.image_data import ImageData
 from pydocx.openxml.vml.shape import Shape
+from pydocx.openxml.vml.textbox import Textbox
 
 __all__ = [
     'ImageData',
     'Shape',
+    'Textbox',
 ]