+++ pydocx/openxml/vml/textbox.py
@@ -6,13 +6,11 @@ from __future__ import (
 )
 
 from pydocx.models import XmlModel, XmlCollection
-from pydocx.openxml.wordprocessing.run import Run
 
 
-class SmartTagRun(XmlModel):
-    XML_TAG = 'smartTag'
+class Textbox(XmlModel):
+    XML_TAG = 'textbox'
 
     children = XmlCollection(
-        Run,
-        'wordprocessing.SmartTagRun',
+        'wordprocessing.TxBxContent',
     )