+++ pydocx/openxml/wordprocessing/textbox_content.py
@@ -6,13 +6,10 @@ from __future__ import (
 )
 
 from pydocx.models import XmlModel, XmlCollection
-from pydocx.openxml.wordprocessing.run import Run
 
 
-class SmartTagRun(XmlModel):
-    XML_TAG = 'smartTag'
-
+class TxBxContent(XmlModel):
+    XML_TAG = 'txbxContent'
     children = XmlCollection(
-        Run,
-        'wordprocessing.SmartTagRun',
+        'wordprocessing.Paragraph',
     )