+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
-**0.9.7**
+**dev**
 
 - Text colors other than black and white are no longer ignored
+- Textboxes have been implemented. We no longer lose the content inside of
+  them.
 
 **0.9.6**