+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-**dev**
+**0.9.7**
 
 - Text colors other than black and white are no longer ignored
 - Textboxes have been implemented. We no longer lose the content inside of